added prover_type to separate deletion and insertion provers
